+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Security
 
 - Don't serve portable objects that are not owned by actors with local accounts.
+- Ignore portable activities from remote servers with local compatible IDs.
 
 ## [4.1.0] - 2025-04-26